% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_fixation_saccade.R
\name{classify_fixation_saccade}
\alias{classify_fixation_saccade}
\title{Title}
\usage{
classify_fixation_saccade(
  data,
  method,
  velocity_threshold = 0.15,
  duration_threshold = 1,
  dispersion_threshold = 0.1,
  estimated_hz = 10,
  raw_output = F,
  verbose = T
)
}
\arguments{
\item{data}{TODO}

\item{method}{TODO}

\item{velocity_threshold}{TODO}
}
\value{
TODO
}
\description{
Title
}
\examples{
TODO
}
