% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_hz.R
\name{estimate_hz}
\alias{estimate_hz}
\title{estimate the fps/hz of the input data}
\usage{
estimate_hz(data)
}
\arguments{
\item{data}{TODO}
}
\value{
a list combining mean_hz, median_hz, and sd_hz TODO
}
\description{
estimate the fps/hz of the input data
}
\examples{
TODO
}
