% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mould_velocity_threshold.R
\name{find_mould_velocity_threshold}
\alias{find_mould_velocity_threshold}
\title{Title}
\usage{
find_mould_velocity_threshold(data, Hz = 10, h = NULL, plot = F)
}
\arguments{
\item{data}{TODO}

\item{h}{the span passing to loess regression}

\item{plot}{TODO}

\item{hz_methods}{which method to estimate hz}
}
\value{
TODO
}
\description{
Title
}
