% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation_and_saccade.R
\name{fixation_and_saccade}
\alias{fixation_and_saccade}
\title{Title}
\usage{
fixation_and_saccade(
  velocities,
  mould_velocity_threshold,
  mould_duration_threshold,
  Hz
)
}
\arguments{
\item{velocities}{TODO}

\item{mould_velocity_threshold}{TODO}

\item{mould_duration_threshold}{TODO}

\item{Hz}{TODO}
}
\value{
TODO
}
\description{
Title
}
