% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_local_maximum.R
\name{get_local_maximum}
\alias{get_local_maximum}
\title{A local maximum}
\usage{
get_local_maximum(velocities, time_interval = 0.1, methods = "delta")
}
\arguments{
\item{time_interval}{in seconds TODO}

\item{methods}{"diff" or "delta" TODO}

\item{data}{a tibble TODO}
}
\value{
a tibble TODO
}
\description{
A local maximum
}
\examples{
get_local_maximum(data) TODO
}
