% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_moved_distance.R
\name{get_moved_distance}
\alias{get_moved_distance}
\title{calculate cumulative moving distances for each experiment, participant, condition, and device}
\usage{
get_moved_distance(data)
}
\arguments{
\item{data}{a tibble Experiment, Participant, Condition, Trial, Device, x, y, z columns}
}
\value{
a tibble with Experiment, Participant, Condition, Trial, Device, and a moved_distance column
}
\description{
calculate cumulative moving distances for each experiment, participant, condition, and device
}
\examples{
get_moved_distance(data) TODO
}
