% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{get_velocity TODO}
\usage{
get_velocity(data, time_interval = 0.1, method = "delta")
}
\arguments{
\item{data}{a tibble TODO}

\item{time_interval}{a number in seconds, used when the method is \code{diff} TODO}

\item{method}{\code{diff}, \code{delta} TODO}
}
\value{
a tibble of velocity samples
}
\description{
get_velocity TODO
}
\examples{
get_velocity(data) TODO

}
