% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rotation.R
\name{get_rotation}
\alias{get_rotation}
\title{Title}
\usage{
get_rotation(data, time_interval = 0)
}
\arguments{
\item{data}{a tibble}

\item{time_interval}{a number}
}
\value{
a tibble
}
\description{
Title
}
\examples{
get_rotation(data)
}
