% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_velocity.R
\name{get_velocity}
\alias{get_velocity}
\title{get_velocity}
\usage{
get_velocity(data, time_interval = 100, methods = "diff")
}
\arguments{
\item{data}{tibble}

\item{time_interval}{a number in milliseconds, used when the methods is \code{diff}}

\item{methods}{\code{diff}, \code{delta}}
}
\value{
a tibble of velocity samples
}
\description{
get_velocity
}
\examples{
get_velocity(data)

}
